Streamline test setup
This is an optional PR, checking to see what folks think.

This PR reduces the boilerplate code with test setup. It uses Go test's `TempDir` method to make temp directories. These are automatically cleaned up, so we don't have to deal with defer and clean up methods. The PR also stores the test object, so it's not necessary to pass around via method arguments.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>